+++ yash-cli/tests/scripted_test/option-y.sh
@@ -35,32 +35,6 @@ printed
 __OUT__
 
 # TODO not yet implemented
-test_OE -e 0 -f 'pipefail on: single command successful pipe' --pipefail
-true
-__IN__
-
-# TODO not yet implemented
-test_OE -e 13 -f 'pipefail on: single command unsuccessful pipe' --pipefail
-(exit 13)
-__IN__
-
-# TODO not yet implemented
-test_OE -e 0 -f 'pipefail on: multi-command successful pipe' --pipefail
-true | true | true | true
-__IN__
-
-# TODO not yet implemented
-test_OE -e 7 -f 'pipefail on: multi-command unsuccessful pipe' --pipefail
-true | exit 2 | true | exit 7 | true | true
-__IN__
-
-# TODO not yet implemented
-test_OE -e 7 -f 'pipefail on: multi-command unsuccessful pipe in subshell' \
-    --pipefail
-(true | exit 2 | true | exit 7 | true | true)
-__IN__
-
-# TODO not yet implemented
 test_oE -f 'traceall on: effect' --traceall
 exec 2>&1
 COMMAND_NOT_FOUND_HANDLER='echo not found $* >&2; HANDLED=1'