+++ yash-env/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.8.1] - Unreleased
 
+### Added
+
+- The `pipefail` shell option (`option::Option::PipeFail`)
+
 ### Changed
 
 - External dependency versions: